Refactor out ArticleViewController into extensions
